Personalize favicons
This commit adds a favicon package generated by RealFaviconGenerator
using the icon "seedling" #1633409, purchased with a royalty-free
license from the Noun Project.

Although unlikely use cases for this blog, to be comprehensive, it adds
support for 'Add to Home Screen' on iOS and Android, pin to desktop on
Windows, and macOS Safari's pinned tabs and Touch Bar features.

Although RealFaviconGenerator recommends storing favicons in the root
directory, I chose to store them in assets/favicons because the
drawbacks listed of not doing so did not seem to outweigh my preference
to keep a cleaner root directory. (See: 'Why do you recommend to put all
files at the root of the web site?' at
https://realfavicongenerator.net/faq)