Add 'Blog' to masthead
When I shared a link to my About page with a friend, it was apparent
that they didn't realize that clicking on "Michael Fryar" in the
masthead would take them to my home page which currently contains my
blog posts. This commit adds a separate link to "Blog" to prevent that
confusion.